Improve Linux Desktop Bootstrap
Use ansible-core instead of ansible.

ansible contains the ansible collections which are not needed for
linux-desktop-base.yml.

Update README.md to run the scripts with `sh` (to avoid make the script executable).

Finally also add vim-scripts in the linux-desktop-base.yml, so that VIM starts without
problems, when using my default .vimrc.local dotfile.