Download raw linux-desktop-bootstrap.sh
Instead of downloading the whole repository, we tell the user to just
download one script: linux-desktop-bootstrap.sh

Also, remove the `-K` (`--ask-become-password`) option from the
ansible-playbook call. sudo was already called, so it should be cached,
we do not need to prompt for the sudo password if it is not needed.