Improve bootstrap line, fix typos in ansible tasks
Piping a shell script from the internet is not the best practice, but as
the script is really simple, it makes life a bit easier.

Create /usr/local/share/keyrings in `linux-desktop-base.yml` at it is
needed nearly in every playbook.

A colon is not allowed in yaml files in ansible playbooks, they must be
enclosed in quotes.

Added the --yes option to gpg, so that the public keyring file can be
overwritten without prompting the user.