linux-desktop-devops.yml: Add htop, iotop und strace
htop: interactive processes viewer Htop is an ncursed-based process
viewer similar to top, but it allows one to scroll the list vertically
and horizontally to see all processes and their full command lines.

iotop is a simple top-like I/O monitor. iotop does for I/O usage what
top(1) does for CPU usage. It watches I/O usage information output by
the Linux kernel and displays a table of current I/O usage by processes
on the system. It is handy for answering the question "Why is the disk
churning so much?".

strace is a system call tracer: i.e. a debugging tool which prints out a
trace of all the system calls made by another process/program.  The
program to be traced need not be recompiled for this, so you can use it
on binaries for which you don't have source.