linux-desktop-extended-base.yml: Add apt-file, ddgr + tasksh
apt-file is a command line tool for searching files contained in
packages. This works offline, but you need to run 'apt update' before.
"User" caches can be configured for apt-file (see "Special
configurations similar to apt-file" in apt-file(1)), but as I hardly use
it, I did not configure it.

ddgr is the command line interface to DuckDuckGo. Calling the
DuckDuckGo Bang "!dfiles" does something similar to apt-file: it
performs a Debian file search, but online via packages.debian.org.

The problem is that the exclamation mark (!) in the shell represents a
historical expansion that needs to be escaped. I configured a zsh alias
to be able to search for files in Debian packages online:

  alias dfiles="ddgr -j \!dfiles"

All of the above was needed because taskwarrior in Debian/bookworm no
longer includes 'tasksh', a shell command that wraps taskwarrior
commands. I needed the above tools to find the appropriate package.