Update role github_install
We want the role to run correctly in check_mode.
It should query the release and try to parse the download URL.
If the download URL of the role is empty, the role should fail, even in
check_mode.

The Github release page uses `name` and `tag_name` to identify the
release. Some repositories use the one, other the other, but `tag_name`
seems to be safe bet. Let's try that.

Also add a description to the role.