linux-desktop-office.yml: Install font related packages
After various problems with some documents and their fonts, I decided to
install more rather than less font packages.

Apart from fonts-crosextra-carlito + fonts-crosextra-caladea,
ttf-mscorefonts-installer is the most useful for me, as this is the
installer for Microsoft TrueType core fonts, which are used in quite a
few documents we receive.