linux-desktop-extended-base.yml: Add libsecret-tools + sort package list alphabetically
libsecret-tools is tool for storing and retrieving GObject passwords
and provides a command line tool `secret-tool` using libsecret to access
the freedesktop.org Secret Service API. This can be used to store and
retrieve passwords for desktop applications.

The tool is usefull in combination with borg + borgmatic to retrieve the
repository password for automatic backups.