linux-desktop-office.yml: Add gucharmap
gucharmap is a Unicode character picker and font browser.
It allows you to browse through all the available Unicode characters and
categories for the installed fonts, and to examine their detailed
properties. It is an easy way to find the character you might only know
by its Unicode name or code point.

It is invoked by the Cinnamon Keyboard layout applet when you click Show
Character Table.