github_install: Use the download URL provided by the releases/latest API
We used a complicated way to find the latest release / tag name.
With it we tried to get the download URL of the latest release.

This is not needed. We can get the download URL directly from the
releases/latest API call.

The actual reason for this fix was, that original jq call to retrieve
github_download_url sometimes returned multiple URLs like this one:

  ❯ jq -r '.[].assets[] | select(.browser_download_url | test(".*1.3.1.*x86_64.deb$")) | .browser_download_url' rustdesk_releases
  https://github.com/rustdesk/rustdesk/releases/download/1.3.1/rustdesk-1.3.1-x86_64.deb
  https://github.com/rustdesk/rustdesk/releases/download/nightly/rustdesk-1.3.1-x86_64.deb

Our new approach should be more robust.